doesn't detect Xeon e5-1650 v3 as a cpu that is better than i5-750 #50

+++ source/Models/BenchmarkData.cs
@@ -10,7 +10,7 @@ namespace SystemChecker.Models
     {
         public string id { get; set; }
         public string name { get; set; }
-        public int mark { get; set; }
+        public float mark { get; set; }
         public int rank { get; set; }
     }
 }
